requirements.txt
requests
python-telegram-bot
langchain
langchain-community